move-only: Move CBlockTreeDB to node/blockstorage
Can be reviewed with --color-moved=dimmed-zebra
